+++ src/util/isMobile.tsx
@@ -1,4 +1,4 @@
 import { isMobile as _isMobile } from 'react-device-detect';
 export const isMobile = (): boolean => {
-    return _isMobile || (localStorage.getItem("__polymorph_mobile_testing") == "true")
+    return _isMobile || (localStorage.getItem("__Workflowish_mobile_testing") == "true")
 }